+++ bpm-engine-impl/src/main/java/io/takari/bpm/form/FormTaskHandler.java
@@ -64,7 +64,7 @@ public class FormTaskHandler implements UserTaskHandler {
 
         formService.create(pk, fId, eventName, fd, options, env);
 
-        return StateHelper.push(state, new CreateEventAction(definitionId, elementId, eventName, null, null, null));
+        return StateHelper.push(state, new CreateEventAction(definitionId, elementId, eventName, null, null, null, null));
     }
 
     private Map<String, Object> getOptions(FormExtension x, Variables vars) {