+++ bpm-engine-impl/src/test/java/io/takari/bpm/EventSchedulerTest.java
@@ -27,7 +27,7 @@ public class EventSchedulerTest {
             @Override
             public Event answer(InvocationOnMock invocation) throws Throwable {
                 UUID id = (UUID)invocation.getArguments()[0];
-                return new Event(id, null, null, null, null, null, true, null);
+                return new Event(id, null, null, null, null, null, true, null, null);
             }
         });